// frontend/src/api/model/getAlertHistoryParams.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DevOps Dashboard API
 * 运维监控仪表盘一期 API 规范（Spec-Driven Development）
 * OpenAPI spec version: 1.0.0
 */

export type GetAlertHistoryParams = {
/**
 * 页码
 */
page?: number;
/**
 * 每页条数
 * @maximum 100
 */
pageSize?: number;
/**
 * 级别筛选（info/warning/critical，空为全部）
 */
level?: string;
};