// frontend/src/api/model/getAlertHistory200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DevOps Dashboard API
 * 运维监控仪表盘一期 API 规范（Spec-Driven Development）
 * OpenAPI spec version: 1.0.0
 */
import type { Alert } from './alert';

export type GetAlertHistory200 = {
  list: Alert[];
  total: number;
  page: number;
  pageSize: number;
};