// frontend/src/api/model/alertLevel.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DevOps Dashboard API
 * 运维监控仪表盘一期 API 规范（Spec-Driven Development）
 * OpenAPI spec version: 1.0.0
 */

export type AlertLevel = typeof AlertLevel[keyof typeof AlertLevel];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AlertLevel = {
  info: 'info',
  warning: 'warning',
  critical: 'critical',
} as const;